+++ bootique-jdbc-junit5/src/main/java/io/bootique/jdbc/junit5/connector/DeleteBuilder.java
@@ -16,51 +16,32 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+package io.bootique.jdbc.junit5.connector;
 
-package io.bootique.jdbc.test;
+import io.bootique.jdbc.junit5.metadata.DbColumnMetadata;
 
-import io.bootique.jdbc.test.jdbc.ExecStatementBuilder;
+/**
+ * @since 2.0.B1
+ */
+public class DeleteBuilder {
 
-public class UpdateSetBuilder {
+    protected final ExecStatementBuilder builder;
 
-    protected ExecStatementBuilder builder;
-    protected int setCount;
-
-    protected UpdateSetBuilder(ExecStatementBuilder builder) {
+    public DeleteBuilder(ExecStatementBuilder builder) {
         this.builder = builder;
     }
 
-    /**
-     * @return the number of updated records.
-     * @since 0.24
-     */
-    public int exec() {
-        return builder.exec();
-    }
-
-
-    public UpdateSetBuilder set(String column, Object value) {
-        return set(column, value, Column.NO_TYPE);
-    }
-
-    public UpdateSetBuilder set(String column, Object value, int valueType) {
-        if (setCount++ > 0) {
-            builder.append(", ");
-        }
-
-        builder.appendIdentifier(column)
-                .append(" = ")
-                .appendBinding(column, valueType, value);
-        return this;
-    }
-
     public WhereBuilder where(String column, Object value) {
-        return where(column, value, Column.NO_TYPE);
+        return where(column, value, DbColumnMetadata.NO_TYPE);
     }
 
     public WhereBuilder where(String column, Object value, int valueType) {
         WhereBuilder where = new WhereBuilder(builder);
         where.and(column, value, valueType);
         return where;
+    }
+
+    public int exec() {
+        return builder.exec();
     }
 }