+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 2.0.B1
 
 * #105 Upgrade Testcontainers to 1.15.0-rc2
+* #106 Test API: Table.delete().where(..)
 
 ## 2.0.M1